test pour la colision
ca avance j'ai encore quelques trucs a regler pour eviter d'avoir un mouvement sacader

+++ src/Display/DisplaySolid.java
@@ -22,7 +22,7 @@ public class DisplaySolid extends JPanel {
     }
 
     public void display() {
-        setBounds(solid.x(), solid.y(), solid.w(), solid.h());
+        setBounds((int) solid.x(), (int) solid.y(), solid.w(), solid.h());
     }
